Remove separate file
